Fix tooltip only showing once
I was trying to write a dynamic tooltip for `UPowerWidget` and noticed the tooltip was only showing up once. I figured out that it was because `self._tooltip_timer` still had a value in `_start_tooltip`, which requires it to be falsey. I have updated `_stop_tooltip` to explicitly set `_tooltip_timer` to `None`.