Add Visualiser widget
Adds a new `Visualiser` widget to provide an audio visualiser in the
bar. The widget relies on `cava` for the audio analysis and offloads
then generation of the image data to a separate script which is not
constrained by qtile's event loop.

It's probably got lots of bugs in but is good enough to merge now!