Fix mirror handling of decorations
Mirrors need to draw decorations separately, particularly when using
PowerLineDecorations as this looks at the next weidget in the bar for
the colouring.

In order to do this, decorations cannot be drawn to the RecordingSurface
otherwise mirrors will draw the same decoration. Instead, the
decorations are now drawn directly to the window surface.

Fixes #136