Reduce documentation for compatibility fixes
Some widgets are just modified because they need changes to work with
qtile-extras (typically with regards to rendering decorations).

These widgets should not be listed as new widgets on the home page nor
do they need full documentation. Instead, a short note should be
provided saying this is a compatibility fix.