Fix attribute error in decorations
Fixes a bug where a widget has multiple decorations, one of which is
grouped and another which is a `PowerLineDecoration`. The bug arises
because the `PowerLineDecoration` does not inherit the `GroupMixin` and,
as a result, does not set the `group` attribute. However, when grouped
widgets checking grouping, they loop over all decorations and check the
`group` attribute. This therefore fails when the loop hits the
`PowerLineDecoration`.

This PR fixes the issue by hardcoding the `group` attribute for
`PowerLineDecoration` objects.

Fixes #248