Add ability to customise position of widget menus
The popup toolkit has some powerful funcitonality for customising the
location of the popup. This PR exposes those methods to exposed commands
for widgets using the popup menu. Clicking on widgets retains the
default behaviour (menu appears by the widget) but it's also possible to
call the command with specific arguments (e.g. binding a key to open
IWD's network list in the middle of the screen).