Fix volume widget bug - unnecessary display
The pulsevolume widget receives a notification from the server when a
new app accesses the sound device. This triggers the callback in the
widget but, as the widget does not check to see if the volume has
changed, the volume is always displayed.

This PR fixes the issue by only displaying the volume when there is a
change.